requirements.txt
Flask==2.3.2
anthropic==24.1.2
pandas==2.0.3
openpyxl==3.1.2
Werkzeug==2.3.6
Pillow==9.5.0
python-dotenv==1.0.0
